Maintain the mapping for more than 1 URL in the localstorage. - rename mappings.js
